hirelens-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
supabase
pdfplumber
python-docx
scikit-learn
nltk
python-jose[cryptography]
sentry-sdk
